moved everything from App.js into a new component Posts.jsx for better readability
